fix: audio fixes
This commit fixes several audio issues discovered in #23.

Way-Shell didn't understand that the wireplumber provided volume in a
cubic scale when a GET to the volume API was made.

We do the appropriate conversion to and from the linear and cubic scales
now.

Additionally, we had some event loops with volume scales. Adjusting
scales could cause an situation where the adjustment now wrote to the
mixer API, and a loop would occur. This is handled by blocking signals.

+++ src/services/wireplumber_service.h
@@ -175,7 +175,7 @@ WirePlumberService *wire_plumber_service_get_global(void);
 
 void wire_plumber_service_set_volume(WirePlumberService *self,
                                      const WirePlumberServiceNode *node,
-                                     float volume);
+                                     double volume);
 
 void wire_plumber_service_volume_up(WirePlumberService *self,
                                     const WirePlumberServiceNode *node);
